fixed missing object label
